fix(leaderboard): correctly counts all consumptions
In older databases it could be that some negative consumption counts are not marked as consumptions, but instead as upgrades. The logic of retrieving the top amount of consumptions was repeated, one of this did not exclude all negative transactions which were not marked as consumptions. As a result, the leadboard would show an incorrect count, limiting access to true leaderboard members.